Split entry tests to meet file size limit
